cum tweaks

+++ include/mystr.h
@@ -5,11 +5,11 @@
 #define MYSTR_H
  
 
-void mystr_Delete (char S[], unsigned int index, unsigned int len);
+void mystr_Delete (char S[], size_t index, size_t len);
 
 
 
-void mystr_Insert (char Res[], char Source[], unsigned int Pos);
+void mystr_Insert (char Res[], char Source[], size_t Pos);
 
 
 
@@ -21,7 +21,7 @@ int mystr_Match (char Source[], char Pattern[]);
 
 
 
-void mystr_Slice (char R[], char S[], unsigned int P, unsigned int L);
+void mystr_Slice (char R[], char S[], size_t P, size_t L);
 
 
 
@@ -29,8 +29,8 @@ void mystr_Subst (char Source[], char OldPattern[], char NewPattern[]);
 
 
 
-void mystr_Position (char Substring[], char Source[], unsigned int at,
-                     unsigned int * pos, int * found);
+void mystr_Position (char Substring[], char Source[], size_t at,
+	size_t * pos, size_t * found);
